Restructure code to perpare for merge
