+++ python/spider-py/src/spider_py/__init__.py
@@ -1,7 +1,7 @@
 """Spider package root."""
 
-from spider.client import chain, group, TaskContext, TaskGraph
-from spider.type import Double, Float, Int8, Int16, Int32, Int64
+from spider_py.client import chain, group, TaskContext, TaskGraph
+from spider_py.type import Double, Float, Int8, Int16, Int32, Int64
 
 __all__ = [
     "Double",
