+++ python/spider-py/src/spider_py/core/data.py
