+++ python/spider-py/src/spider_py/type/tdl_type.py
